refactor: rename types module to Guard and update imports
