+++ packages/pure-parse/docs/guide/guards.md
@@ -1,6 +1,6 @@
 # Guards
 
-Guards are functions that accepts one `unknown` argument and returns a [type predicate](https://www.typescriptlang.org/docs/handbook/advanced-types.html#using-type-predicates):
+A [`Guard`](/api/guards/Guard) are functions that accepts one `unknown` argument and returns a [type predicate](https://www.typescriptlang.org/docs/handbook/advanced-types.html#using-type-predicates):
 
 ```ts
 type Guard<T> = (data) => data is T
@@ -24,10 +24,7 @@ if (isUser(data)) {
 }
 ```
 
-> [!TIP]
-> For a full reference, see the [API documentation on parsers](/api/guards).
-
-## Overview
+## API Reference Overview
 
 PureParse exports two categories of functions related to type guarding.
 
@@ -81,194 +78,4 @@ const isUsers = array(
 ```
 
 > [!TIP]
-> See the [API Reference documentation](/api/guards) for a complete inventory
-
-## Primitives
-
-Primitive types represent [primitive values](https://developer.mozilla.org/en-US/docs/Glossary/Primitive), which are immutable and have no properties:
-
-```ts
-isString('hello') // -> true
-isNumber(42) // -> true
-isBoolean(true) // -> true
-isNull(null) // -> true
-isUndefined(undefined) // -> true
-isBigInt(42n) // -> true
-isSymbol(Symbol()) // -> true
-```
-
-## Equality Checks for Primitive Literals
-
-Primitive literals such as `true`, `false`, `42`, `"hello"`, and `null` are all types _and_ values (depending on the context they appear in). Use the `equalsGuard()` function to create a guard function that compares the data with the strict equality operator (`===`):
-
-```ts
-const isRed = equalsGuard('red')
-isRed('red') // -> true
-isRed('blue') // -> false
-
-const isOne = equalsGuard(1)
-isOne(1) // -> true
-isOne(2) // -> false
-```
-
-To validate a union of literals, use `oneOfGuard`:
-
-```ts
-const isDirection = oneOfGuard(
-  equals('north'),
-  equals('south'),
-  equals('east'),
-  equals('west'),
-)
-isDirection('north') // -> true
-isDirection('east') // -> true
-
-const isDigit = oneOfGuard(
-  equalsGuard(0),
-  equalsGuard(1),
-  equalsGuard(2),
-  equalsGuard(3),
-  equalsGuard(4),
-  equalsGuard(5),
-  equalsGuard(6),
-  equalsGuard(7),
-  equalsGuard(8),
-  equalsGuard(9),
-)
-isDigit(5) // -> true
-isDigit(100) // -> false
-```
-
-## Unions
-
-Unions types—or sum types—represent values that can be one of several types. Use the `oneOfGuard()` function to create a validation function for a union type:
-
-```ts
-const isStringOrNumber = unionGuard(isString, isNumber)
-isStringOrNumber('hello') // -> true
-isStringOrNumber(123) // -> true
-```
-
-Since it is very common to create unions with `null` and `undefined`, there are two helper functions for validating optional and nullable types:
-
-- `undefineableGuard`—for unions with `undefined`
-- `nullableGuard`—for unions with `null`
-
-```ts
-const isUndefineableString = undefineableGuard(isString)
-isOptionalString('hello') // -> true
-isOptionalString(undefined) // -> true
-
-const isNullableString = optionalGuard(isString)
-isNullableString('hello') // -> true
-isNullableString(null) // -> true
-```
-
-## Tuples
-
-Tuples are arrays of fixed length, where each element has a specific type. Use the `tupleGuard()` function to create a guard function for a tuple type:
-
-```ts
-const isCoordinate = tupleGuard([isNumber, isNumber])
-isCoordinate([42, 34]) // -> true
-
-const isTransparentColor = tupleGuard([isString, isNumber])
-isTransparentColor(['#FF0000', 0.5]) // -> true
-```
-
-## Objects
-
-Validate objects with the `objectGuard()` function which takes an object with keys and corresponding validation functions:
-
-```ts
-const isUser = objectGuard({
-  id: isNumber,
-  name: isString,
-})
-isUser({ id: 42, name: 'Alice' }) // -> true
-```
-
-You can nest objects:
-
-```ts
-const isUser = objectGuard({
-  id: isNumber,
-  name: isString,
-  address: objectGuard({
-    country: isString,
-    city: isString,
-    streetAddress: isString,
-    zipCode: isNumber,
-  }),
-})
-```
-
-You can declare optional properties:
-
-```ts
-const isUser = objectGuard({
-  id: isNumber,
-  name: optionalGuard(isString),
-})
-isUser({ id: 42 }) // -> true
-isUser({ id: 42, name: undefined }) // -> true
-isUser({ id: 42, name: 'Jiří' }) // -> true
-```
-
-You can explicitly declare the type of the object and annotate the validation function with the type as a type parameter:
-
-```ts
-type User = {
-  id: number
-  name?: string
-}
-const isUser = objectGuard<User>({
-  id: isNumber,
-  name: optionalGuard(isString),
-})
-```
-
-## Arrays
-
-Arrays are ordered sets of elements of the same type. Use the `arrayGuard()` function to create a validation function for an arrays type:
-
-```ts
-const isBase = oneOf(
-  equalsGuard('A'),
-  equalsGuard('T'),
-  equalsGuard('C'),
-  equalsGuard('G'),
-)
-const isDna = arrayGuard(isBase)
-isDna(['A', 'T', 'A', 'T', 'C', 'G']) // -> true
-```
-
-When explicitly declaring array types, provide type of the item in the array type argument:
-
-```ts
-// Guard<number[]>
-const isNumberArray = arrayGuard<number>(isNumber)
-```
-
-## Tagged/Discriminated Unions
-
-Validate discriminated unions with unions of objects with a common tag property:
-
-```ts
-const isState = unionGuard(
-  objectGuard({
-    tag: equalsGuard('loading'),
-  }),
-  objectGuard({
-    tag: equalsGuard('error'),
-    error: isString,
-  }),
-  objectGuard({
-    tag: equalsGuard('loaded'),
-    message: isString,
-  }),
-)
-isState({ tag: 'loading' }) // -> true
-isState({ tag: 'error', error: 'Failed to load' }) // -> true
-isState({ tag: 'loaded', message: 'Data loaded' }) // -> true
-```
+> For a full reference, see the [API documentation on parsers](/api/guards).
