+++ packages/pure-parse/src/guards/Guard.ts
