+++ packages/pure-parse/src/guards/arrays.ts
@@ -1,4 +1,4 @@
-import { Guard } from './types'
+import { Guard } from './Guard'
 
 /**
  * Validate arrays
